Factor out page-storage, comments & tweaks.
