+++ src/activity-logger/background/log-page-visit.js
@@ -1,5 +1,5 @@
 import { generateVisitDocId, isWorthRemembering } from '..'
-import { reidentifyOrStorePage } from './log-page'
+import { reidentifyOrStorePage } from '../../page-storage/store-page'
 
 
 // Store the visit in PouchDB.
