+++ src/page-storage/index.js
@@ -1,0 +1,14 @@
+import docuri from 'docuri'
+import randomString from '../util/random-string'
+
+export const pageKeyPrefix = 'page/'
+
+export const convertPageDocId = docuri.route(`${pageKeyPrefix}:timestamp/:nonce`)
+
+export function generatePageDocId({timestamp, nonce}={}) {
+    const date = timestamp ? new Date(timestamp) : new Date()
+    return convertPageDocId({
+        timestamp: date.getTime(),
+        nonce: nonce || randomString(),
+    })
+}
